Split config.py into conflicts, build policy, index routing and value parsers
